requirements.txt
glfw == 2.4.0
PyGLM == 2.5.5
PyOpenGL == 3.1.5
PyOpenGL-accelerate == 3.1.5
numpy == 1.21.4